+++ packages/themes/styles/scss/labkey/components/_header.scss
@@ -307,3 +307,8 @@
     background-color: $lk-grid-row-color-hover;
   }
 }
+
+// Used in header of Assay, Sample Type, and Source pages across products
+.page-detail-header-title-padding {
+    padding-left: 5px;
+}
